Experimental Changeling changes pt.1 - Reducing amount of required absorbs
About the PR

Removed the changeling bloodtest exploding. To identify a ling, you will now have to see them perform changeling activities, like a heretic casting spells.
Reduced the amount of DNA and absorbs required as a changeling objective.
Why / Balance

Changelings, as an antagonist, especially with them becoming a team antagonist, have caused issues due to no longer having to fight for victims. Being able to work together encouraged them to go loud especially often, due to their need to take the DNA of about 1 person per 15 minutes, per changeling. The problem is further complicated by the fact the changelings cannot take the DNA of someone that has been absorbed, a way of aquiring DNA that changelings usually prefer due to the evolution points they aquire.

In addition, it is actively encouraged to not stay hidden - when a security officer or a command member randomly demanding a blood test cannot be denied without causing suspicion ( which is usually followed by a wordless stun + cuff if security has been briefed and changelings are known to be on station ), it is much safer to attempt to just kill as many people as possible with your hivemates.

These changes should hopefully increase the viability of not being caught as a changeling without forcing a pacifist playstyle.

There are, of course, potential issues that may arise with the more difficult identification of changelings. That may still however be possible due to the following reasons:

    Changeling bodies not rotting
    Changeling abilities are clearly visible, such as the armblade, EMP scream and other things clearly able to be recognized by command
    In the case of two identical people, command may act the same way as they do with paradox clones - It is much more fun to play trivia than to wordlessly take blood from both people while they wait in the brig!

Technical details
Media
Requirements

    [x ] I have read and am following the Pull Request and Changelog Guidelines.
    [ x] I have added media to this PR or it does not require an ingame showcase.